requirements.txt
langchain==0.3.18
langchain_community==0.3.17
langchain_core==0.3.34
PyYAML==6.0.2
wxauto==3.9.11.17.5
